cms-admin: Extract TipTap content translation logic into its own file
createTipTapRichTextBlock.tsx was growing hard to navigate as translation
support was added on top of the block's existing input/output/preview logic.
Move the translation-specific pieces (externalizing link/child-block data for
the HTML round trip, restoring it, and the HTML-based translateTipTapContentAsync)
into contentTranslation.ts, keeping the generic content-mapping helpers
(mapLinkMarksData, mapCmsBlockNodesData) in the main file since they're also
used by input2State/state2Output/output2State/createPreviewState.
